+++ src/be/nikiroo/utils/serial/CustomSerializer.java
@@ -23,7 +23,7 @@ public abstract class CustomSerializer {
 	public boolean encode(StringBuilder builder, Object value) {
 		int prev = builder.length();
 		String customString = toString(value);
-		builder.append("custom:").append(getType()).append(":");
+		builder.append("custom^").append(getType()).append("^");
 		if (!SerialUtils.encode(builder, customString)) {
 			builder.delete(prev, builder.length());
 			return false;
@@ -37,23 +37,23 @@ public abstract class CustomSerializer {
 	}
 
 	public static boolean isCustom(String encodedValue) {
-		int pos1 = encodedValue.indexOf(':');
-		int pos2 = encodedValue.indexOf(':', pos1 + 1);
+		int pos1 = encodedValue.indexOf('^');
+		int pos2 = encodedValue.indexOf('^', pos1 + 1);
 
-		return pos1 >= 0 && pos2 >= 0 && encodedValue.startsWith("custom:");
+		return pos1 >= 0 && pos2 >= 0 && encodedValue.startsWith("custom^");
 	}
 
 	public static String typeOf(String encodedValue) {
-		int pos1 = encodedValue.indexOf(':');
-		int pos2 = encodedValue.indexOf(':', pos1 + 1);
+		int pos1 = encodedValue.indexOf('^');
+		int pos2 = encodedValue.indexOf('^', pos1 + 1);
 		String type = encodedValue.substring(pos1 + 1, pos2);
 
 		return type;
 	}
 
 	public static String contentOf(String encodedValue) {
-		int pos1 = encodedValue.indexOf(':');
-		int pos2 = encodedValue.indexOf(':', pos1 + 1);
+		int pos1 = encodedValue.indexOf('^');
+		int pos2 = encodedValue.indexOf('^', pos1 + 1);
 		String encodedContent = encodedValue.substring(pos2 + 1);
 
 		return encodedContent;